requirements.txt
# ── Data Ingestion ─────────────────────────────────────────────────────────────
yfinance>=0.2.36
pandas-datareader>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0       # for S&P 500 constituent scraping

# ── Core Data Processing ───────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# ── Backtesting ────────────────────────────────────────────────────────────────
vectorbt>=0.26.1

# ── Performance Analytics ──────────────────────────────────────────────────────
quantstats>=0.0.62
pyfolio-reloaded>=0.9.5

# ── Visualization ──────────────────────────────────────────────────────────────
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0
kaleido>=0.2.1               # plotly static image export

# ── Config & Utilities ─────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0                 # progress bars
tabulate>=0.9.0              # pretty-print tables

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
